v0.7.0c: bugfixes: - fixed reach weapons not working for non-heroes - fixed stealth using integers rather than floats - fixed magical sleep decrementing paralysis without incrementing it

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/Char.java
@@ -491,7 +491,7 @@ public abstract class Char extends Actor {
 		}
 	}
 	
-	public int stealth() {
+	public float stealth() {
 		return 0;
 	}
